+++ test/unit/finances.test.ts
@@ -5,18 +5,30 @@ import { createMockHttpClient, mockMwsFail, mockMwsServiceStatus, parsingError }
 function mockFunctions() {
   /**
    * Mock everything in purify-ts, restore it back to normal,
-   *  except for `enumeration` which will be stubbed
+   *  except for `enumeration` which will be stubbed to accept
    * https://github.com/facebook/jest/issues/936#issuecomment-265074320
    */
   const original = jest.requireActual('purify-ts')
+
   return {
     ...original, // Pass down all the exported objects
-    enumeration: () => {
+    enumeration: <T extends Record<string, string | number>>(enumeration: T) => {
+      const enumValues = Object.values(enumeration)
+
       return original.Codec.custom({
         decode: (input: string | number) => {
-          return original.Right(input)
+          if (typeof input !== 'string' && typeof input !== 'number') {
+            return original.Left(`Expected enum, received ${input}`)
+          }
+
+          const enumIndex = enumValues.indexOf(input)
+
+          return enumIndex !== -1 || input === 'String'
+            ? original.Right((enumValues[enumIndex] as T[keyof T]) || 'String')
+            : original.Left(`Expected enum, received ${input}`)
         },
         encode: original.identity,
+        schema: () => ({ enum: enumValues }),
       })
     },
   }
@@ -67,8 +79,6 @@ describe('finances', () => {
       expect(
         await mockListFinancialEvents.finances.listFinancialEvents(parameters),
       ).toMatchSnapshot()
-
-      jest.clearAllMocks()
     })
 
     it('returns a next token and financial events list if succesful', async () => {